-added ULM diagram -added Schema of database -updated weblayer tests
