+++ src/test/java/PerkManager/WebLayers/SmokeTest.java
@@ -1,21 +1,22 @@
-package anthonymp.SYSC4806_Lab3;
+package PerkManager.WebLayers;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import PerkManager.Controllers.WebPageController;
 import org.junit.jupiter.api.Test;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-/*
+
 @SpringBootTest
 public class SmokeTest {
 
     @Autowired
-    private HomePageController homePageController;
+    private WebPageController webPageController;
 
     //Test if context is creating HomePageController
     @Test
     void contextLoads() {
-        assert (homePageController != null);
+        assert (webPageController != null);
     }
-}*/
+}
