+++ src/test/java/PerkManager/WebLayers/WebMockTest.java
@@ -1,0 +1,34 @@
+package PerkManager.WebLayers;
+
+import static org.hamcrest.Matchers.containsString;
+import static org.mockito.Mockito.when;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+import PerkManager.Controllers.WebPageController;
+import org.junit.jupiter.api.Test;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.web.servlet.MockMvc;
+
+
+@SpringBootTest
+@AutoConfigureMockMvc
+//@WebMvcTest(HomePageController.class)
+public class WebMockTest {
+
+	@Autowired
+	private MockMvc mockMvc;
+
+	@Test
+	public void greetingShouldReturnMessageFromService() throws Exception {
+		//Not Implemented
+    }
+}
