--- src/test/java/anthonymp/SYSC4806_Lab3/GetRequestTests.java
@@ -1,26 +1,0 @@
-package anthonymp.SYSC4806_Lab3;
-/*
-@WebMvcTest(HomePageController.class)
-public class GetRequestTests {
-
-
-    //RestTemplate restTemplate = new RestTemplate();
-
-    @Autowired
-    private MockMvc mockMvc;
-
-    String homePageURL = "http://localhost:8080/greeting";
-    ResponseEntity<String> response = mockMvc.getDispatcherServlet(homePageURL + "", String.class);
-
-
-    @Test
-    public void testGetRequestsTest() {
-        Assertions.assertEquals(response.getStatusCode(), HttpStatus.OK);
-    }
-
-    @Test
-    public void httpHeadersRequestTest() {
-        //HttpHeaders httpHeaders = restTemplate.headForHeaders(homePageURL);
-        //Assertions.assertTrue(httpHeaders.getContentType().includes(MediaType.APPLICATION_JSON));
-    }
-}*/
